Separated init logic from helpers
